kreda na tablah

+++ Flask_bootstrap v1 - samo WEBpage/app.py
@@ -37,8 +37,8 @@ def randomNumberGenerator():
         else:
             statistikaNobena+=1
         ravnotezen=statistikaLeva+statistikaDesna+statistikaNobena
-        leva=statistikaLeva*100/ravnotezen
-        desna=statistikaDesna*100/ravnotezen
+        leva=int(statistikaLeva*100/ravnotezen)
+        desna=int(statistikaDesna*100/ravnotezen)
         number=[number1, number2, leva, desna, (100-leva-desna)] #tale array posljemo preko sock emit na spletno stran
         socketio.emit('newnumber', {'number': number}, namespace='/test')
         socketio.sleep(5) #osvezimo na vsake 5 sec
